Create LICENSE.md
The protocol has been written to facilitate free and open interoperability and is thus released under a minimal restriction, free license. This protocol may be freely used in accordance with  copyright section mentioned below:

Put in simple terms, the terms of using this protocol standards  are: 
      You may use this protocol for any reason – commercial or  non-commercial. You are allowed to modify the protocol to  suit the needs of your implementation. In case you choose to modify the protocol, you MUST submit the changes back to us so that we can incorporate these changes back to the protocol. After all, we wrote this protocol to ensure that IOT devices interoperate from any vendor can interoperate with devices from other vendors!
=====================================
Copyright
=====================================

Copyright (C) Qantom Software Pvt. Ltd. (2019-2020).  All Rights Reserved. RELEASED TO THE PUBLIC DOMAIN AS AN OPEN STANDARD with the following conditions

This document and translations of it may be copied and furnished to others, and derivative works that comment on or otherwise explain it or assist in its implementation may be prepared, copied, published and distributed, in whole or in part, provided that 
     a) The above copyright notice and this paragraph are included on all such copies and derivative works.  However, this document itself may not be modified in any way, such as by removing thecopyright notice or references to Qantom Software Pvt. Ltd or other organizations, involved in this excercise. 
      b) Any modifications made to the protocol specification must be submitted back to Qantom software so that it might be incorporated by Qantom back to the protocol, if appropriate. The reason why Qantom has made this protocol as an open protocol is to ensure that IOT devices are more interoperable and to follow a standard open protocol. We expect a reciprocal from all entities who use this protocol and wish to modify /adapt this protocol to suit their needs. 

The limited permissions granted above are perpetual and will not be revoked by Qantom Software or its successors or assigns. This document and the information contained herein is provided on an "AS IS" basis and Qantom Software Private Limited DISCLAIMS ALL WARRANTIES, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO ANY WARRANTY THAT THE USE OF THE INFORMATION HEREIN WILL NOT INFRINGE ANY RIGHTS OR ANY IMPLIED WARRANTIES OF MERCHANTABILITY OR FITNESS FOR A PARTICULAR PURPOSE.