Add the first copy of the public release
Ver 0.8 of the IONMP (Input Output Node Messaging Protocol) Specification. This is the first  release of the IONMP Specification into the public domain. IONMP is a plug and play protocol that allows sensors and actuators to register with an infrastructure, advertise their identity / capabilities, provides a standard way for sensors to publish their readings and actuators to be controlled by the infrastructure. 
The key features of this protocol are that it makes the participants of the system as plug and play entities and allows run time discovery of device capabilities.